# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Confidential Indexer API",
    "version": "0.1.0",
    "description": "Read-only HTTP API surfacing cleartext balances and transfer history for ERC-7984 confidential tokens. Designed for wallet partners — three distinct null balance reasons (never_shielded / no_decrypt_rights / encrypted_pending) are exposed explicitly so partner UI can render each honestly without conflating them."
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Liveness + indexer state + NATS stream stats",
        "tags": [
          "health"
        ],
        "description": "Single endpoint for ops dashboards: indexer head block, transfer-status counts, and live NATS consumer stats polled from the /jsz monitoring port. `nats` is null when /jsz is unreachable so /health stays healthy under partial outage (alert on null for ops). Configure NATS_MONITORING_URL to override the auto-derived monitoring URL.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "indexer": {
                      "type": "object",
                      "properties": {
                        "lastBlock": {
                          "type": "integer"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      }
                    },
                    "counts": {
                      "type": "object",
                      "properties": {
                        "done": {
                          "type": "integer"
                        },
                        "ready": {
                          "type": "integer"
                        },
                        "noAcl": {
                          "type": "integer"
                        },
                        "failed": {
                          "type": "integer"
                        }
                      }
                    },
                    "nats": {
                      "type": [
                        "null",
                        "object"
                      ],
                      "properties": {
                        "stream": {
                          "type": "string"
                        },
                        "messages": {
                          "type": "integer"
                        },
                        "bytes": {
                          "type": "integer"
                        },
                        "consumers": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string"
                              },
                              "numPending": {
                                "type": "integer"
                              },
                              "numAckPending": {
                                "type": "integer"
                              },
                              "numRedelivered": {
                                "type": "integer"
                              },
                              "numWaiting": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Prometheus scrape endpoint",
        "tags": [
          "health"
        ],
        "description": "Text format (`# HELP / # TYPE` lines + gauges). PG-side gauges always present: confidential_indexer_transfers_total{status}, signers_enabled, last_block. NATS-side gauges (nats_stream_messages, nats_consumer_num_pending, etc.) present only when /jsz reachable. Primary HPA signal is `nats_consumer_num_pending{stream='decrypt-work',consumer='worker-<addr>'}`.",
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/balance/{addr}": {
      "get": {
        "summary": "Current cleartext balance for an address",
        "tags": [
          "balances"
        ],
        "description": "Three distinct null cases: `never_shielded` (no encrypted balance handle exists), `no_decrypt_rights` (handle exists, no signer in this indexer can decrypt), and `encrypted_pending` (handle exists, decrypt queued or in progress). When `amount` is present, `source` distinguishes whether it came from a userDecrypt call or an on-chain AmountDisclosed event.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "addr",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "addr": {
                      "type": "string"
                    },
                    "amount": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "source": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "reason": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "currentHandle": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "updatedAtBlock": {
                      "type": [
                        "null",
                        "integer"
                      ]
                    },
                    "stale": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transfers/{addr}": {
      "get": {
        "summary": "Transfer history for an address",
        "tags": [
          "transfers"
        ],
        "description": "Returns transfers where the address is from or to (or both). `amountCleartext` is null for `no_acl` (we lack decrypt rights), `ready` (queued), or `failed` (all signers exhausted) — `status` and `reason` distinguish each. Pagination uses opaque cursors stable across reorgs; pass `nextCursor` from the previous response as `cursor` to fetch the next page.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false,
            "description": "Opaque pagination cursor from a previous response's nextCursor"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "in",
                "out",
                "both"
              ],
              "default": "both"
            },
            "in": "query",
            "name": "direction",
            "required": false,
            "description": "Filter by transfer direction relative to :addr"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "addr",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "limit",
                    "count",
                    "hasMore"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "block": {
                            "type": "integer"
                          },
                          "logIndex": {
                            "type": "integer"
                          },
                          "txHash": {
                            "type": "string"
                          },
                          "contract": {
                            "type": "string"
                          },
                          "from": {
                            "type": "string"
                          },
                          "to": {
                            "type": "string"
                          },
                          "handle": {
                            "type": "string",
                            "description": "bytes32 ciphertext pointer"
                          },
                          "amountCleartext": {
                            "type": [
                              "null",
                              "string"
                            ],
                            "description": "Decimal string (NUMERIC); null if not decrypted yet"
                          },
                          "source": {
                            "type": [
                              "null",
                              "string"
                            ],
                            "description": "user_decrypt | disclosed | null"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "ready",
                              "done",
                              "no_acl",
                              "failed"
                            ]
                          },
                          "assignedSigner": {
                            "type": [
                              "null",
                              "string"
                            ]
                          },
                          "triedSigners": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "lastError": {
                            "type": [
                              "null",
                              "string"
                            ]
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "reason": {
                            "type": [
                              "null",
                              "string"
                            ],
                            "description": "no_decrypt_rights | encrypted_pending | all_signers_exhausted | null when amountCleartext present"
                          }
                        }
                      }
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "nextCursor": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/operators/{holder}": {
      "get": {
        "summary": "ERC-7984 operator approvals for a holder",
        "tags": [
          "operators"
        ],
        "description": "Lists current operator addresses approved to transfer the holder's confidential tokens. `until` is the approval expiration as a unix-seconds timestamp; operators with `until` in the past are technically expired on chain but kept here for audit (the spec lets clients filter).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "holder",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "holder",
                    "operators"
                  ],
                  "properties": {
                    "holder": {
                      "type": "string"
                    },
                    "operators": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "operator",
                          "until",
                          "setAtBlock"
                        ],
                        "properties": {
                          "operator": {
                            "type": "string"
                          },
                          "until": {
                            "type": "integer",
                            "description": "unix-seconds timestamp; 0 if never approved"
                          },
                          "setAtBlock": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/signers": {
      "post": {
        "summary": "Register or update a signer; backfill no_acl rows now decryptable by it",
        "tags": [
          "admin"
        ],
        "description": "Upserts the signer row (insert or update existing). Runs `backfillForNewSigner` unconditionally — even if NATS is unreachable, PG state stays correct and the worker picks up rows on next message. NATS publishes are best-effort; failure is surfaced as `natsError` in the response, not as a 5xx (PG-only success is still success). The `config` object's shape is backend-specific: `{ privateKeyEnv: 'NAME' }` for `local_eoa`, `{ vaultAccountId, assetId, apiKey, secretPath }` for `fireblocks`, `{ keyId, region }` for `aws_kms`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "addr",
                  "kind",
                  "costRank"
                ],
                "properties": {
                  "addr": {
                    "type": "string",
                    "description": "EIP-55 checksummed Ethereum address; the universal join key"
                  },
                  "kind": {
                    "type": "string",
                    "enum": [
                      "local_eoa",
                      "fireblocks",
                      "aws_kms",
                      "silence_labs",
                      "turnkey"
                    ]
                  },
                  "costRank": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "Lower = preferred. Used by pickSignerForTransfer ORDER BY."
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": true,
                    "default": {},
                    "description": "Provider-specific opaque config"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "addr",
                    "backfilled"
                  ],
                  "properties": {
                    "addr": {
                      "type": "string"
                    },
                    "backfilled": {
                      "type": "integer",
                      "description": "Number of no_acl rows flipped to ready by this signer becoming eligible"
                    },
                    "natsError": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "description": "Non-null if NATS publish of wake-ups failed; PG state is still correct"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/signers/{addr}": {
      "delete": {
        "summary": "Disable a signer; pending rows naturally reassigned via escalator on next failure",
        "tags": [
          "admin"
        ],
        "description": "Sets `enabled=FALSE` on the row. The signer is kept in `app.signers` for audit trail (so future `tried_signers` entries remain resolvable). Rows currently in `status='ready'` with `assigned_signer = this` are NOT eagerly reassigned — the next decrypt attempt's `max_deliver` exhaustion triggers MAX_DELIVERIES, the escalator re-picks via `pickSignerForTransfer` with the disabled signer excluded, and the row moves on. Returns 404 if the signer doesn't exist.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "addr",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "addr",
                    "enabled"
                  ],
                  "properties": {
                    "addr": {
                      "type": "string"
                    },
                    "enabled": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "transfers",
      "description": "Indexed transfers with cleartext amounts and null-reason taxonomy"
    },
    {
      "name": "balances",
      "description": "Per-address cleartext balances with three null-reason variants"
    },
    {
      "name": "operators",
      "description": "ERC-7984 operator approvals (plaintext metadata)"
    },
    {
      "name": "health",
      "description": "Liveness, indexer lag, NATS stream stats, Prometheus scrape"
    },
    {
      "name": "admin",
      "description": "Operator-facing signer management (deploy behind partner gateway)"
    }
  ]
}